ACS-5451: Toggle for direct Event sending - added new bean allowing direct event sending and a toggle (switch) property - refactored EventGenerator logic related with creating and sending events - refactored consolidators - renamed EventConsolidator -> NodeEventConsolidator, and moved common logic to new abstract EventConsolidator
