+++ repository/src/main/resources/alfresco/repository.properties
@@ -1236,6 +1236,8 @@ repo.event2.filter.childAssocTypes=rn:rendition
 repo.event2.filter.users=
 # Topic name
 repo.event2.topic.endpoint=amqp:topic:alfresco.repo.event2
+# Specifies if messages should be enqueued in in-memory queue or send directly to topic
+repo.event2.queue.skip=false
 # Thread pool for async enqueue of repo events
 repo.event2.queue.enqueueThreadPool.priority=1
 repo.event2.queue.enqueueThreadPool.coreSize=8
